audio: Grab Sound.position from audio backend
 * Change `AudioBackend::get_sound_position` to return `f64` to
   match `AudioBackend::get_sound_duration`.
 * Wire up `AudioBackend::get_sound_position` to `Sound.position`.
 * Remove unimplmeneted warning from `Sound.position`.

+++ core/src/backend/audio/mixer.rs
@@ -406,10 +406,10 @@ impl AudioMixer {
     /// Returns the position of a playing sound in milliseconds.
     ///
     ////// Returns `None` if the sound is no longer playing.
-    pub fn get_sound_position(&self, instance: SoundInstanceHandle) -> Option<u32> {
+    pub fn get_sound_position(&self, instance: SoundInstanceHandle) -> Option<f64> {
         let sound_instances = self.sound_instances.lock().unwrap();
         // TODO: Return actual position
-        sound_instances.get(instance).map(|_| 0)
+        sound_instances.get(instance).map(|_| 0.0)
     }
 
     /// Returns the duration of a registered sound in milliseconds.
@@ -714,7 +714,7 @@ macro_rules! impl_audio_mixer_backend {
         }
 
         #[inline]
-        fn get_sound_position(&self, instance: SoundInstanceHandle) -> Option<u32> {
+        fn get_sound_position(&self, instance: SoundInstanceHandle) -> Option<f64> {
             self.$mixer.get_sound_position(instance)
         }
